+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 functions/
 .env
 .DS_Store
+.netlify
