+++ netlify.toml
@@ -1,4 +1,7 @@
 [build]
-  functions = "functions"
   command = "yarn build"
+  functions = "functions"
   publish = "public"
+
+[dev]
+  publish="public"
