+++ src/db/index.js
@@ -1,9 +1,10 @@
 import { Pool } from "pg";
 
-let pgPool;
+let pool;
+let losPool;
 
 async function createPool() {
-	pgPool = new Pool({
+	pool = new Pool({
 		host: process.env.DB_HOST,
 		database: process.env.DB_NAME,
 		user: process.env.DB_USER,
@@ -15,9 +16,30 @@ async function createPool() {
 	});
 }
 
+async function createLosPool() {
+	losPool = new Pool({
+		host: process.env.LOS_DB_HOST,
+		database: process.env.LOS_DB_NAME,
+		user: process.env.LOS_DB_USER,
+		password: process.env.LOS_DB_PASS,
+		port: process.env.LOS_DB_PORT,
+		ssl: {
+			mode: "require"
+		}
+	});
+}
+
 export async function performQuery(text, values) {
-	if (!pgPool) await createPool();
-	const client = await pgPool.connect();
+	if (!pool) await createPool();
+	const client = await pool.connect();
+	const result = await client.query(text, values);
+	client.release();
+	return result.rows;
+}
+
+export async function performLosQuery(text, values) {
+	if (!losPool) await createLosPool();
+	const client = await losPool.connect();
 	const result = await client.query(text, values);
 	client.release();
 	return result.rows;
